backend/requirements.txt
alembic==1.17.1
annotated-types==0.7.0
anyio==4.10.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.13.4
bleach==6.2.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
comm==0.2.3
debugpy==1.8.16
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
dnspython==2.8.0
email-validator==2.3.0
executing==2.2.0
fastapi==0.117.1
fastapi-cli==0.0.13
fastapi-cloud-cli==0.2.1
fastjsonschema==2.21.2
fqdn==1.5.1
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
ipykernel==6.30.1
ipython==9.4.0
ipython_pygments_lexers==1.1.1
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.11.0
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
jupyter-events==0.12.0
jupyter-lsp==2.2.6
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
lark==1.2.2
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.3
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.4.5
notebook_shim==0.2.4
numpy==2.3.2
openai==1.109.1
orjson==3.11.3
packaging==25.0
pandocfilters==1.5.1
parso==0.8.4
platformdirs==4.3.8
pluggy==1.6.0
prometheus_client==0.22.1
prompt_toolkit==3.0.51
psutil==7.0.0
psycopg2-binary==2.9.11
pure_eval==0.2.3
pycparser==2.22
pydantic==2.11.9
pydantic-extra-types==2.10.5
pydantic-settings==2.11.0
pydantic_core==2.33.2
pygame==2.6.1
Pygments==2.19.2
pytest==8.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-json-logger==3.3.0
python-multipart==0.0.20
pywin32==310
pywinpty==3.0.0
PyYAML==6.0.2
pyzmq==27.0.2
referencing==0.36.2
requests==2.32.5
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.1.0
rich-toolkit==0.15.1
rignore==0.6.4
rpds-py==0.27.0
Send2Trash==1.8.3
sentry-sdk==2.39.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.44
stack-data==0.6.3
starlette==0.48.0
terminado==0.18.1
tinycss2==1.4.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typer==0.19.2
types-python-dateutil==2.9.0.20250809
typing-inspection==0.4.1
typing_extensions==4.14.1
ujson==5.11.0
uri-template==1.3.0
urllib3==2.5.0
uvicorn==0.37.0
watchfiles==1.1.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
